# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tunas-spec",
    "description": "Tunas REST API",
    "version": "v0.0.1"
  },
  "servers": [
    {
      "url": "{environment}.localhost:8080",
      "description": "local dev",
      "variables": {
        "environment": {
          "default": "dev",
          "description": "Server environment",
          "enum": [
            "dev",
            "prod",
            "qa"
          ]
        }
      }
    }
  ],
...
  "paths": {
...
    "/cart/{userId}/": {
      "get" : {
        "security": [
          {
            "tunasAuth" : []
          }
        ],
        "tags": ["Cart"],
        "parameters": [{
          "in": "path",
          "name": "userId",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
        ],
        "responses": {
          "200": {
            "description": "successfully get user wishlists",
            "content": {
              "application/json" : {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id" : {
                        "type": "string"
                      },
                      "name" : {
                        "type": "string"
                      },
                      "price" : {
                        "type": "number"
                      },
                      "weight" : {
                        "type": "number"
                      },
                      "category" : {
                        "type": "string"
                      },
                      "perUnit" : {
                        "type": "number"
                      },
                      "imageUrl" : {
                        "type": "string"
                      },
                      "total" : {
                        "type": "number"
                      },
                      "product" : {
                        "type": "object",
                        "properties": {

                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cart/{userId}/{productId}": {
      "post" : {
        "security": [
          {
            "tunasAuth" : []
          }
        ],
        "tags": ["Cart"],
        "parameters": [{
          "in": "path",
          "name": "userId",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success move to wishlist",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/schema200"
                },
                "examples": {
                  "success" : {
                    "value": {
                      "response" : "",
                      "metaData" : {
                        "code" : 201,
                        "message" : "Created"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "tunasAuth" : []
          }
        ],
        "tags": ["Cart"],
        "parameters": [{
          "in": "path",
          "name": "userId",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully remove wishlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/schema200"
                },
                "examples": {
                  "success" : {
                    "value": {
                      "response" : "",
                      "metaData" : {
                        "code" : 200,
                        "message" : "OK"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transaction": {
      "post" : {
        "security": [
          {
            "tunasAuth" : []
          }
        ],
        "tags": ["Transaction"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId" : {
                    "type": "string"
                  },
                  "products": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id" : {
                          "type": "string"
                        },
                        "weight" : {
                          "type": "number"
                        },
                        "total" : {
                          "type": "number"
                        },
                        "perUnit" : {
                          "type": "number"
                        },
                        "price" : {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              },
              "examples": {
                "transaction" : {
                  "value": {
                    "userId": "ac723ce6-11d2-11ec-82a8-0242ac130003",
                    "products": [
                      {
                        "id": "a3c02e8c-11d2-11ec-82a8-0242ac130003",
                        "wight": "10",
                        "price": 18500,
                        "perUnit" : 1200,
                        "total": 1
                      }
                    ]
                  }

                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "order product success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/schema200"
                },
                "examples": {
                  "success" : {
                    "value": {
                      "response": "",
                      "metaData": {
                        "code": 200,
                        "message": "OK"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transaction/{transactionId}": {
      "get" : {
        "security": [
          {
            "tunasAuth" : []
          }
        ],
        "tags": ["Transaction"],
        "parameters": [{
          "in": "path",
          "name": "transactionId",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
        ],
        "responses": {
          "200": {
            "description": "successfully make get transaction by id",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                },
                "examples": {
                  "success" : {
                    "value": {
                      "response": {
                        "transaction": {
                          "id": "f438b52e-996b-11ec-a3ab-173cd0023be6",
                          "totalPrice": 3300,
                          "transactionDate": "2022-03-01T14:29:07.956+00:00",
                          "userName": "admin",
                          "userMobile": null,
                          "userEmail": null
                        },
                        "detailTransaction": [
                          {
                            "name": "Broccoli",
                            "id": "f442c74f-996b-11ec-a3ab-97dafbd1b08d",
                            "imageUrl": null,
                            "product": "a3c02e8c-11d2-11ec-82a8-0242ac130003",
                            "price": 1100,
                            "weight": 1500,
                            "perUnit": 100,
                            "total": 2,
                            "transactionId": "f438b52e-996b-11ec-a3ab-173cd0023be6"
                          }
                        ]
                      },
                      "metaData": {
                        "code": 200,
                        "message": "OK"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "tunasAuth" : []
          }
        ],
        "tags": ["Transaction"],
        "parameters": [{
          "in": "path",
          "name": "transactionId",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
        ],
        "responses": {
          "200": {
            "description": "successfully deleting transaction",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/schema200"
                },
                "examples": {
                  "success" : {
                    "value": {
                      "response": "",
                      "metaData": {
                        "code": 200,
                        "message": "OK"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transaction/user/{userId}" : {
      "get" : {
        "security": [
          {
            "tunasAuth" : []
          }
        ],
        "tags": ["Transaction"],
        "parameters": [{
          "in": "path",
          "name": "userId",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
        ],
        "responses": {
          "200": {
            "description": "admin transaction",
            "content": {
              "application/json" : {
                "schema": {
                  "$ref": "#/components/schemas/transaction"
                },
                "examples": {
                  "admin" : {
                    "value": {
                      "response": [
                        {
                          "transaction": {
                            "id": "f438b52e-996b-11ec-a3ab-173cd0023be6",
                            "totalPrice": 3300,
                            "transactionDate": "2022-03-01T14:29:07.956+00:00",
                            "userName": "admin",
                            "userMobile": null,
                            "userEmail": null
                          },
                          "detailTransaction": [
                            {
                              "name": "Broccoli",
                              "id": "f442c74f-996b-11ec-a3ab-97dafbd1b08d",
                              "imageUrl": null,
                              "product": "a3c02e8c-11d2-11ec-82a8-0242ac130003",
                              "price": 1100,
                              "weight": 1500,
                              "perUnit": 100,
                              "total": 2,
                              "transactionId": "f438b52e-996b-11ec-a3ab-173cd0023be6"
                            }
                          ]
                        }
                      ],
                      "metaData": {
                        "code": 200,
                        "message": "OK"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "transaction": {
        "type": "object",
        "properties": {
          "transaction" : {
            "type": "object",
            "properties": {
              "id" : {
                "type": "string"
              },
              "totalPrice" : {
                "type": "number"
              },
              "transactionDate" : {
                "type": "string"
              },
              "userName" : {
                "type": "string"
              },
              "userMobile" : {
                "type": "string"
              },
              "userEmail" : {
                "type": "string"
              }
            }
          },
          "detailTransaction" : {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name" : {
                  "type": "string"
                },
                "id" : {
                  "type": "string"
                },
                "imageUrl" : {
                  "type": "string"
                },
                "product" : {
                  "type": "string"
                },
                "price" : {
                  "type": "number"
                },
                "weight" : {
                  "type": "number"
                },
                "perUnit" : {
                  "type": "number"
                },
                "total" : {
                  "type": "number"
                },
                "transactionId" : {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "metaData": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "number"
          }
        }
      },
      "schema200": {
        "type": "object",
        "properties": {
          "response": {
            "type": "string"
          },
          "metaData": {
            "$ref": "#/components/schemas/metaData"
          }
        }
      },
...
    },
    "securitySchemes": {
      "tunasAuth" : {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "scheme": "http",
        "bearerFormat": "Bearer"
      }
    }
  }
}